[cmake] Make install python module export source path
The path of the installed python module gets now exported
to all downstream modules. We locally also add the module
source path of the own module to the variable
DUNE_PYTHON_SOURCE_PATH in order to use it e.g. in the
doumentation or in other cases where we need all dune
python dependenices of our dune python module.